requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
scikit-learn>=1.4.0
pandas>=2.1.0
numpy<2.0.0,>=1.24.0
pillow>=10.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
httpx>=0.25.0
openai>=1.3.0
transformers>=4.36.0
torch>=2.6.0
torchvision>=0.17.0
sentence-transformers>=2.2.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
datasets>=2.16.0
opencv-python==4.8.1.78
aiofiles==23.2.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-jose==3.3.0
httpx==0.25.2
email-validator>=2.0.0

# RAG (Retrieval-Augmented Generation) Dependencies
chromadb==0.4.24
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
httpx
python-dotenv